Separate issue class and create an index file
